refactor(core): 重构框架配置管理模块，将 erispulse_config.py 重命名为 _self_config.py 并优化配置结构
