+++ src/ErisPulse/Core/_self_config.py
